Verificação de saída e estar no raio feito e toda vez que sai da rota é enviado uma notificação informando que saiu da rota.

+++ app/src/main/java/com/example/rastreamentoinfantil/model/Geofence.kt
@@ -3,6 +3,6 @@ package com.example.rastreamentoinfantil.model
 data class Geofence(
     var id: String? = null,
     var name: String? = null,
-    var radius: Double,
+    var radius: Float,
     var coordinates: Coordinate
 )
